feat: Improving wingetopt security with more flags and other secure function use
Added warning and linker flags to compile wingetopt more securely in Windows and Linux/unix-like environments as recommended by owasp, msft, and openssf.

In the source, using the _s functions from C11 annex K/MSFT's _s functions for security when available.
Since this code already had it's own warnx implementation, I converted to an enum to choose the message to print to resolve a warning from Clang about a non-const format string. Since it's its own internal implementation and use this was a better choice since the number of warnings it emitted was low and easy to handle with the enum.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>