make: Setting _FORTIFY_SOURCE to 3 only when it does not generate a warning
Overriding the default value of _FORTIFY_SOURCE to 3 in Rockylinux 8 generates a lot of warnings that it will only be treated as level 2.
Rather than have all this compiling noise, we can now detect when to enable this higher level of fortification depending on what the system allows or not.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>